move upload analysis to a shared location
